Cleaned Correspondence
This file is originally located in File Manager on DHBox. I transformed the file to remove unwanted information like lines that didn't describe correspondence and changed the formatting of the index to remove page numbers and the word "to". I added column headers and converted the file into a .csv file to produce an organized table. This was done over two days, May 31 to June 1, and finished at 12:30 pm on June 1.